move tiebreakers test to tiebreakers test file
